requirements.txt
# requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
langchain>=0.0.200
langgraph>=0.0.5
groq>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.28.0
python-dotenv>=1.0.0
langchain_groq
langchain_community
duckduckgo-search